DotCompassPointConverter moved to GiGraph.Dot.Entities.Edges.Enums
